Add LinkRenderer and make sure we always use our Markdown-component
