+++ apps/client/src/components/Markdown/index.js
@@ -1,0 +1,21 @@
+import React from "react";
+import ReactMarkdown from "react-markdown";
+
+import { ListItem } from "@datapunt/asc-ui";
+import ListRenderer from "./renderers/ListRenderer";
+import BlockRenderer from "./renderers/BlockRenderer";
+import LinkRenderer from "./renderers/LinkRenderer";
+import Visual from "../Visual";
+
+export default ({ source }) => (
+  <ReactMarkdown
+    source={source}
+    renderers={{
+      paragraph: BlockRenderer,
+      list: ListRenderer,
+      listItem: ListItem,
+      image: Visual,
+      link: LinkRenderer,
+    }}
+  />
+);
