+++ apps/client/src/components/Markdown/renderers/BlockRendererStyles.js
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+import { Paragraph } from "@datapunt/asc-ui";
+
+// IE11 Fix
+export const StyledParagraph = styled(Paragraph)`
+  flex-shrink: 0;
+`;
