+++ apps/client/src/components/Markdown/renderers/ListRenderer.js
@@ -1,0 +1,4 @@
+import React from "react";
+import { List } from "@datapunt/asc-ui";
+
+export default ({ children }) => <List variant="bullet">{children}</List>;
